+++ connectors/golden-tables/src/main/resources/golden/parquet-decimal-dictionaries-v1/.part-00000-92f97f0b-304f-4587-9d25-088cb386fa64-c000.snappy.parquet.crc
