+++ connectors/golden-tables/src/main/resources/golden/parquet-decimal-dictionaries-v1/_delta_log/00000000000000000000.json
@@ -1,0 +1,4 @@
+{"commitInfo":{"timestamp":1691113099049,"operation":"WRITE","operationParameters":{"mode":"ErrorIfExists","partitionBy":"[]"},"isolationLevel":"Serializable","isBlindAppend":true,"operationMetrics":{"numFiles":"1","numOutputRows":"1000000","numOutputBytes":"4571289"},"engineInfo":"Apache-Spark/3.4.0 Delta-Lake/3.0.0-SNAPSHOT","txnId":"62d25bc1-8193-4fb8-89ad-eae3dc17617a"}}
+{"metaData":{"id":"testId","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col1\",\"type\":\"decimal(9,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col2\",\"type\":\"decimal(12,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col3\",\"type\":\"decimal(25,0)\",\"nullable\":true,\"metadata\":{}}]}","partitionColumns":[],"configuration":{},"createdTime":1691113089928}}
+{"protocol":{"minReaderVersion":1,"minWriterVersion":2}}
+{"add":{"path":"part-00000-92f97f0b-304f-4587-9d25-088cb386fa64-c000.snappy.parquet","partitionValues":{},"size":4571289,"modificationTime":1691113098000,"dataChange":true,"stats":"{\"numRecords\":1000000,\"minValues\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0},\"maxValues\":{\"id\":999999,\"col1\":4,\"col2\":5,\"col3\":1},\"nullCount\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0}}"}}
