+++ connectors/golden-tables/src/main/resources/golden/parquet-decimal-dictionaries-v2/_delta_log/00000000000000000000.json
@@ -1,0 +1,4 @@
+{"commitInfo":{"timestamp":1691113113890,"operation":"WRITE","operationParameters":{"mode":"ErrorIfExists","partitionBy":"[]"},"isolationLevel":"Serializable","isBlindAppend":true,"operationMetrics":{"numFiles":"1","numOutputRows":"1000000","numOutputBytes":"60952"},"engineInfo":"Apache-Spark/3.4.0 Delta-Lake/3.0.0-SNAPSHOT","txnId":"9b3ef6e5-cdfd-4890-9a4c-21fc81ffa0f9"}}
+{"metaData":{"id":"testId","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col1\",\"type\":\"decimal(9,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col2\",\"type\":\"decimal(12,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col3\",\"type\":\"decimal(25,0)\",\"nullable\":true,\"metadata\":{}}]}","partitionColumns":[],"configuration":{},"createdTime":1691113109835}}
+{"protocol":{"minReaderVersion":1,"minWriterVersion":2}}
+{"add":{"path":"part-00000-2509b8ef-98ac-42da-98ee-9d2c58ac6031-c000.snappy.parquet","partitionValues":{},"size":60952,"modificationTime":1691113113000,"dataChange":true,"stats":"{\"numRecords\":1000000,\"minValues\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0},\"maxValues\":{\"id\":999999,\"col1\":4,\"col2\":5,\"col3\":1},\"nullCount\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0}}"}}
