+++ connectors/golden-tables/src/main/resources/golden/parquet-decimal-dictionaries/_delta_log/00000000000000000000.json
@@ -1,4 +1,4 @@
-{"commitInfo":{"timestamp":1690695606750,"operation":"WRITE","operationParameters":{"mode":"ErrorIfExists","partitionBy":"[]"},"isolationLevel":"Serializable","isBlindAppend":true,"operationMetrics":{"numFiles":"1","numOutputRows":"1000000","numOutputBytes":"4571289"},"engineInfo":"Apache-Spark/3.4.0 Delta-Lake/3.0.0-SNAPSHOT","txnId":"f9810035-0d46-4f24-84eb-74d7fb3c0b88"}}
-{"metaData":{"id":"testId","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col1\",\"type\":\"decimal(9,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col2\",\"type\":\"decimal(12,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col3\",\"type\":\"decimal(25,0)\",\"nullable\":true,\"metadata\":{}}]}","partitionColumns":[],"configuration":{},"createdTime":1690695598862}}
+{"commitInfo":{"timestamp":1691112043897,"operation":"WRITE","operationParameters":{"mode":"ErrorIfExists","partitionBy":"[]"},"isolationLevel":"Serializable","isBlindAppend":true,"operationMetrics":{"numFiles":"1","numOutputRows":"1000000","numOutputBytes":"60952"},"engineInfo":"Apache-Spark/3.4.0 Delta-Lake/3.0.0-SNAPSHOT","txnId":"8ebef15f-98d0-43ed-aa26-e1f9f1de7f4e"}}
+{"metaData":{"id":"testId","format":{"provider":"parquet","options":{}},"schemaString":"{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col1\",\"type\":\"decimal(9,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col2\",\"type\":\"decimal(12,0)\",\"nullable\":true,\"metadata\":{}},{\"name\":\"col3\",\"type\":\"decimal(25,0)\",\"nullable\":true,\"metadata\":{}}]}","partitionColumns":[],"configuration":{},"createdTime":1691112034932}}
 {"protocol":{"minReaderVersion":1,"minWriterVersion":2}}
-{"add":{"path":"part-00000-c49ddf63-8eef-4144-b454-d42b527fe8f0-c000.snappy.parquet","partitionValues":{},"size":4571289,"modificationTime":1690695606000,"dataChange":true,"stats":"{\"numRecords\":1000000,\"minValues\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0},\"maxValues\":{\"id\":999999,\"col1\":4,\"col2\":5,\"col3\":1},\"nullCount\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0}}"}}
+{"add":{"path":"part-00000-60b8c840-c0d4-428e-9005-89f02233be85-c000.snappy.parquet","partitionValues":{},"size":60952,"modificationTime":1691112043000,"dataChange":true,"stats":"{\"numRecords\":1000000,\"minValues\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0},\"maxValues\":{\"id\":999999,\"col1\":4,\"col2\":5,\"col3\":1},\"nullCount\":{\"id\":0,\"col1\":0,\"col2\":0,\"col3\":0}}"}}
