+++ kernel/kernel-api/src/main/java/io/delta/kernel/expressions/Literal.java
@@ -139,15 +139,14 @@ public final class Literal extends LeafExpression
     /**
      * @return a {@link Literal} with data type {@link DecimalType}
      */
-    // TODO: reconsider behavior here w.r.t. precision/scale for instantiation & evaluation
-    public static Literal of(BigDecimal value, DecimalType dataType)
+    public static Literal of(BigDecimal value, int precision, int scale)
     {
         // throws an error if rounding is required to set the specified scale
-        BigDecimal valueToStore = value.setScale(dataType.getScale());
-        checkArgument(valueToStore.precision() <= dataType.getPrecision(), String.format(
-                "Decimal precision=%s for decimal %s exceeds max precision for data type %s",
-                valueToStore.precision(), valueToStore, dataType));
-        return new Literal(valueToStore, dataType);
+        BigDecimal valueToStore = value.setScale(scale);
+        checkArgument(valueToStore.precision() <= precision, String.format(
+                "Decimal precision=%s for decimal %s exceeds max precision %s",
+                valueToStore.precision(), valueToStore, precision));
+        return new Literal(valueToStore, new DecimalType(precision, scale));
     }
 
     /**
